Feature: Minimap Toggles for adventures (#92)
